Change number of stored profiles to uint8 to allow 255 maximum

+++ contracts/Aavegotchi/libraries/LibWearablesConfig.sol
@@ -10,8 +10,8 @@ library LibWearablesConfig {
     /// @notice Returns the next wearables config id for that gotchi given that owner
     /// @param _owner The owner of the gotchi
     /// @param _tokenId The tokenId of the gotchi
-    /// @return The next free wearables config id
-    function _getNextWearablesConfigId(address _owner, uint256 _tokenId) internal view returns (uint256 nextWearablesConfigId) {
+    /// @return nextWearablesConfigId The next free wearables config id
+    function _getNextWearablesConfigId(address _owner, uint256 _tokenId) internal view returns (uint8 nextWearablesConfigId) {
         AppStorage storage s = LibAppStorage.diamondStorage();
         // slots start at 0 so slotsUsed is always the next config id
         nextWearablesConfigId = s.ownerGotchiSlotsUsed[_owner][_tokenId];
@@ -22,8 +22,8 @@ library LibWearablesConfig {
     /// @param _owner The owner of the gotchi
     /// @param _tokenId The tokenId of the gotchi
     /// @param _wearablesConfigId The wearables config id
-    /// @return True if the wearables config exists false otherwise
-    function _wearablesConfigExists(address _owner, uint256 _tokenId, uint256 _wearablesConfigId) internal view returns (bool exists) {
+    /// @return exists True if the wearables config exists false otherwise
+    function _wearablesConfigExists(address _owner, uint256 _tokenId, uint8 _wearablesConfigId) internal view returns (bool exists) {
         AppStorage storage s = LibAppStorage.diamondStorage();
         // slots start at 0 so slots used should always be greater by 1 than the last config id
         exists = (s.ownerGotchiSlotsUsed[_owner][_tokenId] > _wearablesConfigId);
@@ -31,7 +31,7 @@ library LibWearablesConfig {
 
     /// @notice Checks if a wearables configuration consist of valid wearables and are for the correct slot
     /// @param _wearablesToStore The wearables to store
-    /// @return True if the wearables configuration is valid and false otherwise
+    /// @return valid True if the wearables configuration is valid and false otherwise
     function _checkValidWearables(uint256[EQUIPPED_WEARABLE_SLOTS] memory _wearablesToStore) internal view returns (bool valid) {
         bool valid = true;
         AppStorage storage s = LibAppStorage.diamondStorage();
